+++ sunray/src/driver/SerialRobotDriver.h
@@ -99,7 +99,7 @@ class SerialMotorDriver: public MotorDriver {
     SerialMotorDriver(SerialRobotDriver &sr);
     void begin() override;
     void run() override;
-    void setMotorPwm(int leftPwm, int rightPwm, int mowPwm, float rightSpeed, float leftSpeed) override;
+    void setMotorPwm(int leftPwm, int rightPwm, int mowPwm, int rightSpeed, int leftSpeed) override;
     void getMotorFaults(bool &leftFault, bool &rightFault, bool &mowFault) override;
     void resetMotorFaults()  override;
     void getMotorCurrent(float &leftCurrent, float &rightCurrent, float &mowCurrent) override;
